requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
motor>=3.3.0
pyyaml>=6.0.1
aiohttp>=3.9.0
pydantic>=2.4.0
python-dotenv>=1.0.0